+++ src/Persistence/MassTransit.NHibernateIntegration/MassTransit.NHibernateIntegration.csproj
@@ -14,7 +14,7 @@
 
   <ItemGroup>
         <PackageReference Include="Iesi.Collections" Version="4.0.4" />
-        <PackageReference Include="NHibernate" Version="5.3.3" />
+        <PackageReference Include="NHibernate" Version="5.3.5" />
         <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
         <ProjectReference Include="..\..\MassTransit\MassTransit.csproj" />
   </ItemGroup>
