+++ src/Scheduling/MassTransit.QuartzIntegration/MassTransit.QuartzIntegration.csproj
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-      <PackageReference Include="Quartz" Version="3.2.2" />
+      <PackageReference Include="Quartz" Version="3.2.3" />
       <ProjectReference Include="..\..\MassTransit\MassTransit.csproj" />
       <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0" PrivateAssets="All" />
   </ItemGroup>
