+++ tests/MassTransit.DapperIntegration.Tests/MassTransit.DapperIntegration.Tests.csproj
@@ -7,7 +7,7 @@
   <ItemGroup>
     <PackageReference Include="Appveyor.TestLogger" Version="2.0.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.17.0"/>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1"/>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.0"/>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.5"/>
   </ItemGroup>
